Refactor models.py into internals/.
