+++ framework/basehandlers_test.py
@@ -26,7 +26,7 @@ import werkzeug.exceptions  # Flask HTTP stuff.
 from google.appengine.api import users
 
 from framework import basehandlers
-import models
+from internals import models
 import settings
