+++ internals/fetchmetrics.py
@@ -20,15 +20,12 @@ import datetime
 import json
 import logging
 from xml.dom import minidom
-
-# Appengine imports.
-from framework import ramcache
 import requests
 
-# File imports.
 from framework import basehandlers
+from framework import ramcache
 from framework import utils
-import models
+from internals import models
 import settings
