+++ pages/featuredetail.py
@@ -22,7 +22,7 @@ import logging
 import settings
 from framework import basehandlers
 from pages import guideforms
-import models
+from internals import models
 from internals import processes
